+++ qt_ui/windows/mission/flight/waypoints/QFlightWaypointItem.py
@@ -1,4 +1,4 @@
-from PySide2.QtGui import QStandardItem
+from PySide2.QtGui import QStandardItem, Qt
 
 from game.ato.flightwaypoint import FlightWaypoint
 
@@ -6,6 +6,7 @@ from game.ato.flightwaypoint import FlightWaypoint
 class QWaypointItem(QStandardItem):
     def __init__(self, point: FlightWaypoint, number):
         super(QWaypointItem, self).__init__()
+        self.setData(point, Qt.UserRole)
         self.number = number
         self.setText("{:<16}".format(point.pretty_name))
         self.setEditable(True)